Serious update
Adds a function that gets the class for your character. Adds a function that prioritizes your highest rolls to abilities that are relevant to the class you have been given. Adds a function that gives your race. Adds a function that gives racial modifiers to your rolled ability scores. Adds a function that gets your classes saving throw base scores and base attack bonus. Adds a function that creates a dictionary with a key:value pair of the skill:ability score. Edited the main function to run everything properly